+++ src/richie/plugins/lti_consumer/defaults.py
@@ -1,0 +1,4 @@
+"""Default settings for Richie's LTI consumer plugin."""
+from django.conf import settings
+
+LTI_CONSUMER_CONFIGURATION = getattr(settings, "LTI_CONSUMER_CONFIGURATION", [])